+++ variables.tf
@@ -262,12 +262,6 @@ variable "network_policy" {
   default     = null
 }
 
-variable "enable_confidential_nodes" {
-  type        = bool
-  description = "(Optional) Whether to enable Confidential Nodes for this cluster."
-  default     = false
-}
-
 variable "rbac_security_identity_group" {
   description = "(Optional) The name of the RBAC security identity group for use with Google security groups in Kubernetes RBAC. Group name must be in format 'gke-security-groups@yourdomain.com'."
   type        = string